# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AegisPulse Clinical Deterioration Radar API
  version: 1.0.0
  description: |
    Production-oriented backend API for the AegisPulse Continuous Patient Deterioration Radar.
    Provides deterministic clinical surveillance, early warning scoring (Subbe MEWS, Singer qSOFA),
    explainability reasoning, unified patient timeline analytics, and ward simulation controls.
servers:
  - url: http://localhost:3000
    description: Local Development / Test Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/wards/{wardId}:
    get:
      tags: [Wards & Beds]
      summary: Get ward details and bed inventory
      parameters:
        - name: wardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ward details
        '404':
          description: Ward not found
  /api/v1/wards/{wardId}/overview:
    get:
      tags: [Wards & Beds]
      summary: Ward overview with bed occupancy and patient summary
      parameters:
        - name: wardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ward overview
...
  /api/v1/patients:
    get:
      tags: [Patients]
      summary: List patients
      parameters:
        - name: wardId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of patients
  /api/v1/patients/{patientId}:
    get:
      tags: [Patients]
      summary: Get patient summary and latest observation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient summary
...
  /api/v1/patients/{patientId}/timeline:
    get:
      tags: [Timeline]
      summary: Chronological unified timeline stream
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: integer
        - name: until
          in: query
          schema:
            type: integer
        - name: types
          in: query
          description: Comma-separated list of event types
          schema:
            type: string
        - name: trustedOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Chronological event stream
  /api/v1/patients/{patientId}/timeline/events:
    post:
      tags: [Timeline]
      summary: Record timeline event (manual observation, nurse visit)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimelineEventCreateRequest'
      responses:
        '201':
          description: Event recorded
  /api/v1/patients/{patientId}/timeline/changes:
    get:
      tags: [Timeline]
      summary: "Answers: What changed during the last 4 hours?"
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: hours
          in: query
          schema:
            type: number
            default: 4
      responses:
        '200':
          description: Structured deltas and narrative summary
  /api/v1/patients/{patientId}/timeline/priority-rise:
    get:
      tags: [Timeline]
      summary: "Answers: What caused the patient's priority to rise?"
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: hours
          in: query
          schema:
            type: number
            default: 4
      responses:
        '200':
          description: Priority rise attribution factors and trigger events
  /api/v1/patients/{patientId}/timeline/last-manual-assessment:
    get:
      tags: [Timeline]
      summary: "Answers: When was the patient last manually assessed?"
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Last assessment timestamp, elapsed time, assessor, overdue status
  /api/v1/patients/{patientId}/timeline/trusted-measurements:
    get:
      tags: [Timeline]
      summary: "Answers: Which measurements were trusted?"
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: integer
        - name: vitalType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Segregated trusted and untrusted measurements with suppression reasons
...
  /api/v1/simulation/scenarios/run:
    post:
      tags: [Simulation]
      summary: Switch active simulation scenario (Admin / System required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenarioId]
              properties:
                scenarioId:
                  type: string
                  enum:
                    - NORMAL_SHIFT
                    - SINGLE_PATIENT_DETERIORATION
                    - FALSE_ALARM_SCENARIO
                    - SIGNAL_FAILURE_SCENARIO
                    - MULTIPLE_PATIENT_SCENARIO
      responses:
        '200':
          description: Scenario activated
        '403':
          description: Insufficient privileges
  /api/v1/simulation/tick:
    post:
      tags: [Simulation]
      summary: Advance simulation clock by N seconds (Admin / System required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seconds:
                  type: integer
                  default: 60
      responses:
        '200':
          description: Clock advanced
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    TimelineEventCreateRequest:
      type: object
      required: [eventType, title, description]
      properties:
        eventType:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        source:
          type: string
        isTrusted:
          type: boolean
        data:
          type: object